requirements.txt
Django==1.11.4
djangorestframework==3.6.3
drf-nested-routers==0.90.0
pytz==2017.2
